[tracker] Reduce memory usage for php script server (tracker_fullsync.php)
git-svn-id: https://forge.indepnet.net/svn/tracker/branches/0.71/trunk@1330 349b9182-4a13-0410-896f-e5e9767dd1b3

+++ scripts/tracker_fullsync.php
@@ -124,16 +124,19 @@ if(isset($_GET['update_device_process'])) {
             $ArrayListType[] = $device->infos->type;
             $ArrayListAgentProcess[] = $data['process_number'];
          }
+         unset($xml);
+         unset($device);
       }
    }
+   
    if (isset($ArrayListDevice)) {
       plugin_tracker_UpdateDeviceBySNMP_startprocess($ArrayListDevice,$fields["process_id"],$xml_auth_rep,$ArrayListType,$ArrayListAgentProcess);
    }
-
-   foreach ( $xml_file as $num=>$filename ) {
-      unlink($filename);
+   if ((isset($xml_file)) AND (!empty($xml_file))) {
+      foreach ( $xml_file as $num=>$filename ) {
+         unlink($filename);
+      }
    }
-
    // Create connections between switchs
    $tmpc = new PluginTrackerTmpconnections;
    $tmpc->WireInterSwitchs();
